split up server systems
