+++ packages/instanceserver/src/ServerHostNetworkSystem.tsx
@@ -26,35 +26,8 @@ import { Engine } from '@etherealengine/engine/src/ecs/classes/Engine'
 import { defineSystem } from '@etherealengine/engine/src/ecs/functions/SystemFunctions'
 import { NetworkPeerFunctions } from '@etherealengine/engine/src/networking/functions/NetworkPeerFunctions'
 import { updatePeers } from '@etherealengine/engine/src/networking/systems/OutgoingActionSystem'
-import React, { useEffect } from 'react'
 
-import { DataChannelType } from '@etherealengine/common/src/interfaces/DataChannelType'
-import { NetworkState } from '@etherealengine/engine/src/networking/NetworkState'
-import { NetworkTopics } from '@etherealengine/engine/src/networking/classes/Network'
-import { DataChannelRegistryState } from '@etherealengine/engine/src/networking/systems/DataChannelRegistry'
-import {
-  MediasoupDataConsumerActions,
-  MediasoupDataProducerActions
-} from '@etherealengine/engine/src/networking/systems/MediasoupDataProducerConsumerState'
-import {
-  MediaConsumerActions,
-  MediaProducerActions
-} from '@etherealengine/engine/src/networking/systems/MediasoupMediaProducerConsumerState'
-import { MediasoupTransportActions } from '@etherealengine/engine/src/networking/systems/MediasoupTransportState'
-import { UserID } from '@etherealengine/engine/src/schemas/user/user.schema'
-import { defineActionQueue, getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
 import { SocketWebRTCServerNetwork } from './SocketWebRTCServerFunctions'
-import {
-  createOutgoingDataProducer,
-  handleConsumeData,
-  handleConsumerSetLayers,
-  handleProduceData,
-  handleRequestConsumer,
-  handleRequestProducer,
-  handleWebRtcTransportClose,
-  handleWebRtcTransportConnect,
-  handleWebRtcTransportCreate
-} from './WebRTCFunctions'
 
 export async function validateNetworkObjects(network: SocketWebRTCServerNetwork): Promise<void> {
   for (const [peerID, client] of network.peers) {
@@ -66,93 +39,14 @@ export async function validateNetworkObjects(network: SocketWebRTCServerNetwork)
   }
 }
 
-const requestConsumerActionQueue = defineActionQueue(MediaConsumerActions.requestConsumer.matches)
-const consumerLayersActionQueue = defineActionQueue(MediaConsumerActions.consumerLayers.matches)
-const requestProducerActionQueue = defineActionQueue(MediaProducerActions.requestProducer.matches)
-
-const dataRequestProducerActionQueue = defineActionQueue(MediasoupDataProducerActions.requestProducer.matches)
-const dataRequestConsumerActionQueue = defineActionQueue(MediasoupDataConsumerActions.requestConsumer.matches)
-
-const requestTransportActionQueue = defineActionQueue(MediasoupTransportActions.requestTransport.matches)
-const requestTransportConnectActionQueue = defineActionQueue(MediasoupTransportActions.requestTransportConnect.matches)
-const transportCloseActionQueue = defineActionQueue(MediasoupTransportActions.transportClosed.matches)
-
 const execute = () => {
-  for (const action of requestConsumerActionQueue()) {
-    handleRequestConsumer(action)
-  }
-  for (const action of consumerLayersActionQueue()) {
-    handleConsumerSetLayers(action)
-  }
-  for (const action of requestProducerActionQueue()) {
-    handleRequestProducer(action)
-  }
-
-  for (const action of dataRequestProducerActionQueue()) {
-    handleProduceData(action)
-  }
-  for (const action of dataRequestConsumerActionQueue()) {
-    handleConsumeData(action)
-  }
-
-  for (const action of requestTransportActionQueue()) {
-    handleWebRtcTransportCreate(action)
-  }
-  for (const action of requestTransportConnectActionQueue()) {
-    handleWebRtcTransportConnect(action)
-  }
-  for (const action of transportCloseActionQueue()) {
-    handleWebRtcTransportClose(action)
-  }
-
   const worldNetwork = Engine.instance.worldNetwork as SocketWebRTCServerNetwork
   if (worldNetwork) {
     if (worldNetwork.isHosting) validateNetworkObjects(worldNetwork)
   }
 }
 
-export const DataChannel = (props: { networkID: UserID; dataChannelType: DataChannelType }) => {
-  const { networkID, dataChannelType } = props
-
-  useEffect(() => {
-    const network = getState(NetworkState).networks[networkID] as SocketWebRTCServerNetwork
-    createOutgoingDataProducer(network, dataChannelType)
-
-    return () => {
-      // todo - cleanup
-    }
-  }, [])
-
-  return null
-}
-
-const NetworkReactor = (props: { networkID: UserID }) => {
-  const { networkID } = props
-  const dataChannelRegistry = useHookstate(getMutableState(DataChannelRegistryState))
-  return (
-    <>
-      {dataChannelRegistry.keys.map((dataChannelType) => (
-        <DataChannel key={dataChannelType} networkID={networkID} dataChannelType={dataChannelType as DataChannelType} />
-      ))}
-    </>
-  )
-}
-
-export const reactor = () => {
-  const networkIDs = Object.entries(useHookstate(getMutableState(NetworkState).networks).value)
-    .filter(([networkID, network]) => network.topic === NetworkTopics.world)
-    .map(([networkID, network]) => networkID)
-  return (
-    <>
-      {networkIDs.map((hostId: UserID) => (
-        <NetworkReactor key={hostId} networkID={hostId} />
-      ))}
-    </>
-  )
-}
-
 export const ServerHostNetworkSystem = defineSystem({
-  uuid: 'ee.engine.ServerHostNetworkSystem',
-  execute,
-  reactor
+  uuid: 'ee.instanceserver.ServerHostNetworkSystem',
+  execute
 })
